Remove thematic areas
