+++ app/scripts/components/home/value-propostion.tsx
