+++ app/scripts/utils/veda-data.ts
@@ -1,110 +1,48 @@
-import { useEffect, useMemo, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useParams } from 'react-router';
-import vedaThematics, {
-  thematics,
-  discoveries,
-  datasets,
-  VedaDatum,
-  VedaThematicListItem
-} from 'veda';
+import { discoveries, datasets } from 'veda';
 
 import { MDXContent, MDXModule } from 'mdx/types';
 import { S_IDLE, S_LOADING, S_SUCCEEDED } from './status';
 
+/**
+ * List with the meta information of all datasets.
+ */
 export const allDatasetsProps = Object.values(datasets).map((d) => d!.data);
 
 /**
- * Interface for the Thematic area data and the related discoveries and datasets
+ * List with the meta information of all discoveries.
  */
-export type ThematicItemFull = VedaDatum<VedaThematicListItem> | null;
-
-/**
- * Returns the data for the thematic are taking the url parameter into account.
- * @returns Object
- */
-export function useThematicArea(): ThematicItemFull {
-  const { thematicId, discoveryId, datasetId } = useParams();
-
-  // If there's only one thematic area, the app is setup with only one thematic
-  // area, and therefore return the first one.
-  let tId = vedaThematics.length === 1 ? vedaThematics[0].id : thematicId;
-
-  if (!tId) {
-    const discovery = discoveries[discoveryId ?? ''];
-    const dataset = datasets[datasetId ?? ''];
-    if (discovery) {
-      tId = discovery.data.thematics[0];
-    } else if (dataset) {
-      tId = dataset.data.thematics[0];
-    }
-  }
-
-  return useMemo(() => {
-    if (!tId) return null;
-
-    const relData = vedaThematics.find((d) => d.id === tId);
-    // Relationship data between thematics and datasets/discoveries.
-    const thematic = thematics[tId];
-    if (!relData || !thematic) return null;
-
-    return {
-      ...thematic,
-      data: {
-        ...relData,
-        ...thematic.data
-      }
-    };
-  }, [tId]);
-}
+export const allDiscoveriesProps = Object.values(discoveries).map(
+  (d) => d!.data
+);
 
 /**
  * Returns the meta data for a discovery taking into account the url parameters.
- * If the discovery does not belong to the thematic area being viewed, null is
- * returned.
+ * If the discovery does not exist, null is returned.
  * @returns Object
  */
-export function useThematicAreaDiscovery() {
-  const thematic = useThematicArea();
+export function useDiscovery() {
   const { discoveryId } = useParams();
 
   const discovery = discoveries[discoveryId ?? ''];
 
-  // Stop if the discovery doesn't exist or if it doesn't belong to this
-  // thematic area.
-  if (
-    !thematic ||
-    !discovery ||
-    !discovery.data.thematics.includes(thematic.data.id)
-  ) {
-    return null;
-  }
-
-  return discovery;
+  // Stop if the discovery doesn't exist.
+  return discovery ?? null;
 }
 
 /**
  * Returns the meta data for a dataset taking into account the url parameters.
- * If the dataset does not belong to the thematic area being viewed, null is
- * returned.
+ * If the dataset does not exist, null is returned.
  * @returns Object
  */
-export function useThematicAreaDataset() {
-  const thematic = useThematicArea();
+export function useDataset() {
   const { datasetId } = useParams();
 
   const dataset = datasets[datasetId ?? ''];
 
-  // Stop if the datasets doesn't exist or if it doesn't belong to this
-  // thematic area.
-  if (
-    !thematic ||
-    !dataset ||
-    !dataset.data.thematics.includes(thematic.data.id)
-  ) {
-    return null;
-  }
-
-  return dataset;
+  // Stop if the datasets doesn't exist.
+  return dataset ?? null;
 }
 
 type MdxPageLoadResult =
