--- mock/thematic/agriculture.thematic.mdx
@@ -1,32 +1,0 @@
----
-id: agriculture
-name: 'Agriculture'
-description: "Sed sed lectus vitae odio vestibulum mattis. Integer iaculis nisl lectus, vel aliquet nulla imperdiet in."
-media:
-  src: ::file ./agriculture--cover.jpg
-  alt: 'Maisernte'
-  author:
-    name: no one cares
-    url: https://unsplash.com/photos/l_5MJnbrmrs
-
-about:
-  title: Observing Agriculture
-  description: Lorem ipsum dolor sit amet, consectetur adipiscing elit vitae ornare lectus.
----
-<Block>
-<Prose>
-## Lorem ipsum dolor sit
-
-Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut vitae ornare lectus, ac accumsan erat. Cras eget eleifend ligula. Curabitur ac nisi tempor, molestie lorem vitae, consectetur turpis. Etiam hendrerit et sapien ac interdum.
-
-Phasellus nec magna tristique, imperdiet mauris hendrerit, posuere sem. Vestibulum in lobortis tortor.
-
-Vestibulum justo ante, bibendum at vehicula sit amet, dignissim sit amet augue. Nunc luctus orci mi, sed pretium nibh consequat sit amet. In congue metus a ullamcorper sollicitudin.
-
-### Maecenas non eros
-
-Cras eget eleifend ligula. Curabitur ac nisi tempor, molestie lorem vitae, consectetur turpis. Etiam hendrerit et sapien ac interdum.
-
-Vestibulum justo ante, bibendum at vehicula sit amet, dignissim sit amet augue. Nunc luctus orci mi, sed pretium nibh consequat sit amet. In congue metus a ullamcorper sollicitudin.
-</Prose>
-</Block>
