--- mock/thematic/covid-19.thematic.mdx
@@ -1,29 +1,0 @@
----
-id: covid-19
-name: COVID-19
-description: As communities around the world have changed their behavior in response to the spread of COVID-19, NASA satellites have observed changes in the environment.
-media:
-  src: ::file ./covid-19--cover.jpg
-  alt: 3D rendering of the novel coronavirus observed under the microscope
-  author:
-    name: Fusion Medical Animation
-    url: https://unsplash.com/photos/EAgGqOiDDMg
-
-about:
-  title: Observing COVID-19
-  description: This experimental dashboard reflects a rapid response to COVID-19 that is currently underway and will continue to evolve as more data becomes available.
----
-<Block>
-<Prose>
-## Perceived changes on our planet
-As the world moved indoors to shelter from the global pandemic sparked by the novel coronavirus, we could perceive changes on our planet. The sky seemed a little bluer, the air a little fresher, the animals in our yards more abundant.
-
-The National Aeronautics and Space Administration's (NASA) continuous and sometimes near-real-time measurements of Earth allow for understanding both the systems changes themselves and the potential impact on economies and society during the pandemic – and as the world slowly returns to operations.
-
-This dashboard features data collected and analyzed by NASA. Information about Earth systems is gathered by a fleet of powerful global Earth-observing satellites, instruments aboard the International Space Station, from airborne science campaigns, and via ground observations. With this data we have been able to monitor some of those changes, and this is allowing us to track and compare these changes over time.
-
-NASA will further expand the analysis of the data sets presented here in collaboration with the European Space Agency (ESA) and the Japan Aerospace Exploration Agency (JAXA). ESA and JAXA have created their own online data portals to provide an even richer picture of what is happening on our home planet during this time of crisis. A fourth, tri-agency dashboard will combine the unique data of each agency for more comparisons and deeper understandings now and over time.
-
-To learn more about NASA Earth Science, go to [nasa.gov/earth](https://nasa.gov/earth).
-</Prose>
-</Block>
