--- mock/thematic/void.thematic.mdx
@@ -1,20 +1,0 @@
----
-id: void
-name: Void area
-description: This thematic area is going to be empty. No datatsets or discoveries
-media:
-  src: ::file ./void-placeholder.jpg
-  alt: Light streaks pointing to a persons silhouette
-  author:
-    name: Alexis Fauvet
-    url: https://unsplash.com/photos/L-3oTJhmsW4
-
-about:
-  title: The void
-  description: Emptyness through and through
----
-<Block>
-<Prose>
-Into the void we go...
-</Prose>
-</Block>
