+++ mock/veda.config.js
@@ -1,5 +1,4 @@
 module.exports = {
-  thematics: './thematic/*.thematic.mdx',
   datasets: './datasets/*.data.mdx',
   discoveries: './discoveries/*.discoveries.mdx',
