+++ parcel-resolver-veda/index.d.ts
@@ -111,7 +111,7 @@ declare module 'veda' {
    * editors can curate contents per each category with their ids
    */
   export interface RelatedContentData {
-    type: 'dataset' | 'discovery' | 'thematic';
+    type: 'dataset' | 'discovery';
     id: string;
     thematic?: string;
   }
@@ -152,25 +152,6 @@ declare module 'veda' {
     media?: Media;
     thematics: string[];
     related?: RelatedContentData[];
-  }
-
-  // ///////////////////////////////////////////////////////////////////////////
-  //  Thematic areas                                                          //
-  // ///////////////////////////////////////////////////////////////////////////
-
-  /**
-   * Data structure for the Thematics frontmatter.
-   */
-  export interface ThematicData {
-    id: string;
-    name: string;
-    description: string;
-    media?: Media;
-    related?: RelatedContentData[];
-    about?: {
-      title: string
-      description: string
-    }
   }
 
   // ///////////////////////////////////////////////////////////////////////////
@@ -215,31 +196,6 @@ declare module 'veda' {
    * Object with all the veda discoveries keyed by the discovery id.
    */
   export const discoveries: VedaData<DiscoveryData>;
-
-  /**
-   * Named exports: thematics.
-   * Object with all the veda thematics keyed by the thematic id.
-   */
-  export const thematics: VedaData<ThematicData>;
-
-  export interface VedaThematicListItem extends ThematicData {
-    /**
-     * Datasets that are related to this thematic area.
-     */
-    datasets: DatasetData[];
-    /**
-     * Discoveries that are related to this thematic area.
-     */
-    discoveries: DiscoveryData[];
-  }
-
-  /**
-   * The default export is a list of all the thematic areas with their
-   * respective datasets and discoveries. It contains no MDX content, just the
-   * frontmatter data.
-   */
-  const _default: VedaThematicListItem[];
-  export default _default;
 
   export type PageOverrides = 'aboutContent' | 'sandbox-override';
   /**
